Move KafkaJsonDecoder code to pinot-json (#7021)
* Move KafkaJsonDecoder code to pinot-json and keep KafkaJsonDecoder class for backward compatability

* Add license header

* Add deprecated annotation

* remove Stream from class name
